microservices/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
mysql-connector-python>=8.2.0
pydantic>=2.5.0
httpx>=0.25.0
requests>=2.31.0
